+++ lib/domain/models/coordinates.dart
@@ -1,10 +1,10 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 class Coordinates {
-  int x;
-  int y;
-  int height;
-  int width;
-  int degree;
+  double x;
+  double y;
+  double height;
+  double width;
+  double degree;
   
   Coordinates({
     required this.x,
